+++ 04_entryForm/src/javascript/checkStatus/checkColorFormStatus.js
@@ -9,7 +9,7 @@ export const checkColorFormStatus = () => {
   let status;
 
   field.addEventListener("change", () => {
-    if (field.value !== "") {
+    if (field.value) {
       status = true;
     } else {
       status = false;
