+++ 04_entryForm/src/javascript/checkStatus/checkMileageFormStatus.js
@@ -15,7 +15,7 @@ export const checkMileageFormStatus = () => {
   let status;
 
   field.addEventListener("change", () => {
-    if (field.value !== "") {
+    if (field.value) {
       status = true;
     } else {
       status = false;
