+++ 04_entryForm/src/javascript/constants.js
@@ -1,0 +1,174 @@
+export const MAKERLIST = [
+  {
+    value: "nissan",
+    label: "日産",
+    type: "domestic"
+  },
+  {
+    value: "toyota",
+    label: "トヨタ",
+    type: "domestic"
+  },
+  {
+    value: "matsuda",
+    label: "マツダ",
+    type: "domestic"
+  },
+  {
+    value: "volkswagen",
+    label: "フォルクスワーゲン",
+    type: "overseas"
+  }
+];
+
+export const CARLIST = [
+  {
+    value: "serena",
+    label: "セレナ",
+    maker: "nissan"
+  },
+  {
+    value: "aria",
+    label: "アリア",
+    maker: "nissan"
+  },
+  {
+    value: "sakura",
+    label: "サクラ",
+    maker: "nissan"
+  },
+  {
+    value: "note",
+    label: "ノート",
+    maker: "nissan"
+  },
+  {
+    value: "aqua",
+    label: "アクア",
+    maker: "toyota"
+  },
+  {
+    value: "golf",
+    label: "ゴルフ",
+    maker: "volkswagen"
+  }
+];
+
+export const GRADELIST = [
+  {
+    maker: "toyota",
+    car: "aqua",
+    grade: "G",
+    year: 2021
+  },
+  {
+    maker: "toyota",
+    car: "aqua",
+    grade: "AQUA MARINE",
+    year: 2020
+  },
+  {
+    maker: "nissan",
+    car: "serena",
+    grade: "X",
+    year: 2022
+  },
+  {
+    maker: "nissan",
+    car: "serena",
+    grade: "ハイウェイスターＶ",
+    year: 2022
+  },
+  {
+    maker: "nissan",
+    car: "serena",
+    grade: "オーテック",
+    year: 2022
+  }
+]
+
+export const MILEAGE = [
+  {
+    value: "under10000",
+    label: "1万キロ未満"
+  },
+  {
+    value: "10000over",
+    label: "1万キロ以上"
+  },
+  {
+    value: "20000over",
+    label: "2万キロ以上"
+  },
+  {
+    value: "30000over",
+    label: "3万キロ以上"
+  },
+  {
+    value: "40000over",
+    label: "4万キロ以上"
+  },
+  {
+    value: "50000over",
+    label: "5万キロ以上"
+  },
+  {
+    value: "60000over",
+    label: "6万キロ以上"
+  },
+  {
+    value: "70000over",
+    label: "7万キロ以上"
+  },
+  {
+    value: "80000over",
+    label: "8万キロ以上"
+  },
+  {
+    value: "90000over",
+    label: "9万キロ以上"
+  },
+  {
+    value: "100000over",
+    label: "10万キロ以上"
+  }
+];
+
+export const COLORLIST = [
+  {
+    value: "white",
+    label: "白"
+  },
+  {
+    value: "black",
+    label: "黒"
+  },
+  {
+    value: "gray",
+    label: "グレー"
+  },
+  {
+    value: "silver",
+    label: "シルバー"
+  },
+  {
+    value: "red",
+    label: "赤"
+  },
+  {
+    value: "yellow",
+    label: "黄色"
+  },
+  {
+    value: "blue",
+    label: "青色"
+  },
+  {
+    value: "green",
+    label: "緑色"
+  },
+  {
+    value: "other",
+    label: "その他"
+  }
+];
