--- 04_entryForm/src/javascript/constraints.js
@@ -1,55 +1,0 @@
-export const MAKERLIST = [
-  {
-    value: "nissan",
-    label: "日産",
-    type: "domestic"
-  },
-  {
-    value: "toyota",
-    label: "トヨタ",
-    type: "domestic"
-  },
-  {
-    value: "matsuda",
-    label: "マツダ",
-    type: "domestic"
-  },
-  {
-    value: "volkswagen",
-    label: "フォルクスワーゲン",
-    type: "overseas"
-  }
-];
-
-export const GRADELIST = [
-  {
-    maker: "toyota",
-    car: "aqua",
-    grade: "G",
-    year: 2021
-  },
-  {
-    maker: "toyota",
-    car: "aqua",
-    grade: "AQUA MARINE",
-    year: 2020
-  },
-  {
-    maker: "nissan",
-    car: "serena",
-    grade: "X",
-    year: 2022
-  },
-  {
-    maker: "nissan",
-    car: "serena",
-    grade: "ハイウェイスターＶ",
-    year: 2022
-  },
-  {
-    maker: "nissan",
-    car: "serena",
-    grade: "オーテック",
-    year: 2022
-  }
-]
