+++ 04_entryForm/src/javascript/createSelectItem/createYearSelectItem.js
@@ -1,13 +1,13 @@
 import { createOption } from "../createOption.js"
 
 export const createYearSelectItem = () => {
-  const basicYear = 2023;
+  const nowYear = new Date().getFullYear();
   const yearRange = 15;
 
   const modelYearSelectMenu = document.querySelector("[data-js='modelYearSelectMenu']");
 
   for (let i = 0; i < yearRange; i++) {
-    const year = basicYear - i;
+    const year = nowYear - i;
 
     const item = {
       value: year,
